+++ src/infrastructure/three/grids/index.ts
@@ -1,6 +1,0 @@
-export * from './IGridStyle';
-export * from './GridFactory';
-export * from './SquareWireGrid';
-export * from './HexagonalWireGrid';
-export * from './HexagonalGlassGrid';
-export * from './StoneTileGrid';